% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Función1.R
\name{Verd.Posit}
\alias{Verd.Posit}
\title{Función para filtrar los datos de Scalpel}
\usage{
Verd.Posit(Archivo, CovRatio, NReads)
}
\arguments{
\item{df}{Un txt de annovar, el covratio y el numero de reads}
}
\value{
Ya lo haré \code{df}
}
\description{
Función para filtrar los datos de Scalpel
}
\examples{
Verd.Posit(Archivo="annovar.txt", CovRatio=0.25, NReads=20)
}
